+++ include/RevHart.h
@@ -1,7 +1,7 @@
 //
 // _RevHart_h_
 //
-// Copyright (C) 2017-2024 Tactical Computing Laboratories, LLC
+// Copyright (C) 2017-2025 Tactical Computing Laboratories, LLC
 // All Rights Reserved
 // contact@tactcomplabs.com
 //
@@ -20,7 +20,7 @@ namespace SST::RevCPU {
 
 class RevHart {
   ///< RevHart: Id for the Hart (0,1,2,3,etc)
-  uint32_t ID{};
+  uint16_t ID{};
 
   ///< RevHart: State management object when a Hart is executing a system call
   EcallState Ecall{};
@@ -41,7 +41,7 @@ class RevHart {
 public:
   ///< RevHart: Constructor
   RevHart(
-    uint32_t                                                          ID,
+    uint16_t                                                          ID,
     const std::shared_ptr<std::unordered_multimap<uint64_t, MemReq>>& LSQueue,
     std::function<void( const MemReq& )>                              MarkLoadCompleteFunc
   )
